feat: Add PlotSquared integration
This commit adds support for PlotSquared integration in the BoundaryGuard
project. It includes changes to the IntegrationType enum and the
BoundaryGuard class to include PlotSquared. Also, a new file,
PlotSquaredIntg.java, has been added to implement the integration.

The commit also involves removal of the CustomType.java file. The pom.xml file
has been updated to include dependencies for PlotSquared.